refactor: rename tab to be from 'Distribution' to 'Example'.
This is because we can use a tabBox within a tab to create two overarching tabs, one with dummy data and another with users own data, with the sub-tabs for both of these being for 'preference distribution' and 'preference allocation'. Splitting the tabs up in this way makes more sense.