feat: add section to present raw preference data for viewing.
This is so users can see the preferences made by people within the app, rather than having to go outside the app to see it. This will improve the UX.

It points to an object called 'present_data_preference' which has not been created yet. This means the app won't necessarily run/show the data until we create it in the next commit.