feat: add logic to shut down R session when closing the app.
This is mainly for developers who want to work on the app and restart it when closing the app and making changes to it. Also have renamed the 'Report - Distributions' tab to 'Report - Example' so it is consistent with the naming across the piece. Lastly, added text about the user being able to feed their own data into the app.