fix: adapt iterative preference function to work for vertical data.
This is so we can encourage users to format their data correctly and ensure that the example dataset we are using for the app is consistent with the data format we are encouraging users to abide by. The result of this is we make the code between the 'Report - Example' and 'Report - Allocations' tab consistent with each other.